Allow overriding the favicon
The favicon is a pretty much a static resource, so it is appropriate to allow to override it statically, i.e. via an environment variable.

To override the favicon, set the following environment variable: APP_BRANDING_FAVICON

Not sure if being context-relative does matter here, I could not find any way that this application would be deployed on an application server (e.g. as a WAR), so I *think* ignoring servlet context paths is okay here, but do correct me here if it does matter.