Support redirect urls with existing query parameters
We currently run into an error because the `onboardProcessResult` query param is just appended to the redirect url given to the middlerware, which results in an invalid url e.g. "https://example.org/?test=test?onboardProcessResult=SUCCESS"

Refactor the query param to use UriComponentsBuilder, which parses the URL correctly and only appends a question mark if needed.