Add agrirouter-middleware service to docker-compose.yml
The docker-compose.yml of the agrirouter-middleware-local has been extended to include the configuration of the agrirouter-middleware service. A new section, namely 'agrirouter-middleware', is added for defining the middleware service. It includes settings such as the image, platform, environment variables, ports, network, and dependencies. This will enable enhanced local development and testing by allowing the middleware service to be easily started and stopped through Docker Compose.