Refactor InfoController version method to use Callable
The version response method in the InfoController class has been refactored to return a Callable object. This is to take advantage of Spring MVC's ability to handle Callable, providing a simple way to achieve asynchronous behavior thus enhancing the application performance. It ensures ongoing requests to the version endpoint are processed efficiently without impairing or stalling other operations.