Remove task executor and handle requests synchronously
The task executor used for handling asynchronous tasks was removed from GlobalBeanConfiguration.java and the adjustments were made in SecuredOnboardProcessService.java to deal with requests synchronously which was previously handled by the executor. The initiative behind this change is to improve error management and increase the visibility of the system's behavior. Configured request timeouts and a new filter to log request durations were also added for better management of server resources and identifying slow requests respectively.