"Added redirect delay feature for agrirouter
Incorporated redirect delay into the agrirouter middleware in order to perform redirects smoothly. A new configuration parameter 'redirect-delay' was introduced in the application.yml file to hold the delay time, which is fetched via @Value annotation inside CallbackController.java. Additionally, the redirectDelay value is now passed as a query parameter in the redirect URL and used in the landing_page.html file to delay the redirect via JavaScript. This promotes greater control and customization of redirect behavior."