Update OpenAPI-specification details
Updated 'swagger/openapi.json' to reflect latest changes to API. Increased version number to 8.11.0. Simplified and unified response-conventions for 'agrirouter© callback'. A large number of monitoring, statistics and related operations have been removed. Significantly reduced clutter in the JSON schema. Seemingly RPC-related protobuf constructs have been removed. This reduction reflected a streamlining and refocussing of our middleware. It was done to provide a simpler and clearer interface to the users.