Implement actuator external credentials configuration
Added configuration for external credentials for the actuator in TenantService.java. This allows users to set their own credentials to use the actuator endpoints instead of the generated ones. Updates have also been made to the application.yml and the additional spring configuration metadata, plus the creation of the application-actuator-credentials.yml file to handle these configurations. Instructions on how to use the external credentials have been added to the functional documentation.